refactor: 이미지와 관련된 클래스 패키지 분리
