+++ src/main/java/nextstep/courses/domain/image/SessionImage.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.image;
 
 public class SessionImage {
     private final Long id;
