+++ src/test/java/nextstep/courses/domain/ImageSizeTest.java
@@ -3,6 +3,7 @@ package nextstep.courses.domain;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import nextstep.courses.domain.image.ImageSize;
 import org.junit.jupiter.api.Test;
 
 class ImageSizeTest {
